+++ contracts/libraries/ErrorsLib.sol
@@ -108,6 +108,7 @@ library ErrorsLib {
 
     // ClaimTopicsRegistry Errors
     error ClaimTopicAlreadyExists();
+    error InvalidIdentityType();
 
     // IdentityRegistry Errors
     error EligibilityChecksDisabledAlready();